+++ Specs/Scene/GltfLoaderSpec.js
@@ -2003,6 +2003,9 @@ describe(
           VertexAttributeSemantic.TEXCOORD,
           0
         );
+        var positionQuantization = positionAttribute.quantization;
+        var normalQuantization = normalAttribute.quantization;
+        var texcoordQuantization = texcoordAttribute.quantization;
 
         var indices = primitive.indices;
 
@@ -2036,6 +2039,28 @@ describe(
         expect(positionAttribute.buffer).toBeDefined();
         expect(positionAttribute.byteOffset).toBe(0);
         expect(positionAttribute.byteStride).toBeUndefined();
+        expect(positionQuantization.octEncoded).toBe(false);
+        expect(positionQuantization.normalizationRange).toEqual(
+          new Cartesian3(2047, 2047, 2047)
+        );
+        expect(positionQuantization.quantizedVolumeOffset).toEqual(
+          new Cartesian3(
+            -69.29850006103516,
+            9.929369926452637,
+            -61.32819747924805
+          )
+        );
+        expect(positionQuantization.quantizedVolumeDimensions).toEqual(
+          new Cartesian3(
+            165.4783935546875,
+            165.4783935546875,
+            165.4783935546875
+          )
+        );
+        expect(positionQuantization.componentDatatype).toBe(
+          ComponentDatatype.UNSIGNED_SHORT
+        );
+        expect(positionQuantization.type).toBe(AttributeType.VEC3);
 
         expect(normalAttribute.name).toBe("NORMAL");
         expect(normalAttribute.semantic).toBe(VertexAttributeSemantic.NORMAL);
@@ -2063,6 +2088,14 @@ describe(
         expect(normalAttribute.buffer).toBeDefined();
         expect(normalAttribute.byteOffset).toBe(0);
         expect(normalAttribute.byteStride).toBeUndefined();
+        expect(normalQuantization.octEncoded).toBe(true);
+        expect(normalQuantization.normalizationRange).toBe(255);
+        expect(normalQuantization.quantizedVolumeOffset).toBeUndefined();
+        expect(normalQuantization.quantizedVolumeDimensions).toBeUndefined();
+        expect(normalQuantization.componentDatatype).toBe(
+          ComponentDatatype.UNSIGNED_BYTE
+        );
+        expect(normalQuantization.type).toBe(AttributeType.VEC2);
 
         expect(texcoordAttribute.name).toBe("TEXCOORD_0");
         expect(texcoordAttribute.semantic).toBe(
@@ -2086,6 +2119,20 @@ describe(
         expect(texcoordAttribute.buffer).toBeDefined();
         expect(texcoordAttribute.byteOffset).toBe(0);
         expect(texcoordAttribute.byteStride).toBeUndefined();
+        expect(texcoordQuantization.octEncoded).toBe(false);
+        expect(texcoordQuantization.normalizationRange).toEqual(
+          new Cartesian2(1023, 1023)
+        );
+        expect(texcoordQuantization.quantizedVolumeOffset).toEqual(
+          new Cartesian2(0.026409000158309937, 0.01996302604675293)
+        );
+        expect(texcoordQuantization.quantizedVolumeDimensions).toEqual(
+          new Cartesian2(0.9600739479064941, 0.9600739479064941)
+        );
+        expect(texcoordQuantization.componentDatatype).toBe(
+          ComponentDatatype.UNSIGNED_SHORT
+        );
+        expect(texcoordQuantization.type).toBe(AttributeType.VEC2);
 
         expect(indices.indexDatatype).toBe(IndexDatatype.UNSIGNED_SHORT);
         expect(indices.count).toBe(12636);
@@ -2095,9 +2142,9 @@ describe(
         expect(positionAttribute.buffer).not.toBe(normalAttribute.buffer);
         expect(positionAttribute.buffer).not.toBe(texcoordAttribute.buffer);
 
-        expect(positionAttribute.buffer.sizeInBytes).toBe(28788);
-        expect(normalAttribute.buffer.sizeInBytes).toBe(28788);
-        expect(texcoordAttribute.buffer.sizeInBytes).toBe(19192);
+        expect(positionAttribute.buffer.sizeInBytes).toBe(14394);
+        expect(normalAttribute.buffer.sizeInBytes).toBe(4798);
+        expect(texcoordAttribute.buffer.sizeInBytes).toBe(9596);
       });
     });